New trait config is convincing
It flows nicely.

Removed intermediate "tag" and made it the field type metadata.
It is split from the complete Metadata with offset, so handling the
offset can be made different later (first variants...).

Now only two traits are needed, one for Field operations, and one for defining initializers.

Slices are not for any T, but all T must be set at creation from an iterator with known length.
If using [MaybeUninit<T>], initializing with uninit() can be used.
Thus all use cases covered.

This version drops the idea to share lengths in the header, too complicated.