Deref compound metadata
Expanded trait to support manipulation of a tuple of fields.

This makes use of the newly introduced GAT to define the *tuple of
references* in an ergonomic way.
It could probably be done without GAT but at the cost of one trait per
deref impl (if lifetime deduction works at all).
Not tested in an impl though.