Improve relation between Metadata and Layout
Added documentation.
While doing that, it was noticed that Layout and field descriptor (metadata) are redundant.
Layout can be computed from metadata.
Thus the Initializer trait only needs to return Metadata, and layout can be accessed with the Field trait from the metadata only.
This removes the needs to have duplication of Layout definition, with SAFETY docstrings indicating that they must match.

In the case of Unsized<dyn T> this means getting Layout from vtable.
A godbolt test indicates that devirtualization works during allocate, so this is not a problem.
Drop will have to call the vtable for field drop anyway.

Sadly this does not work as well for FieldSequence.
Allocation in initialize creates both final layout and metadata (with offsets).
FieldSequence::layout does the loop of Layout::extend too but discards offsets.
It is simpler for now to keep them duplicated. FIXME later.